Move Roles to network::config
